change logic of payment

+++ VNPay/Models/PaymentRequest.cs
@@ -16,7 +16,7 @@ namespace VNPay.Models
         /// Mã tham chiếu giao dịch (Transaction reference). Đây là mã số duy nhất dùng để xác định giao dịch.
         /// Lưu ý: Giá trị này bắt buộc và cần đảm bảo không bị trùng lặp giữa các giao dịch.
         /// </summary>
-        public int PaymentId { get; set; }
+        public long PaymentId { get; set; }
 
         /// <summary>
         /// Thông tin mô tả nội dung thanh toán, không dấu và không bao gồm các ký tự đặt biệt.
